requirements.txt
flask
python-dotenv
werkzeug
openai
twilio
pytesseract
pillow
